+++ dev_server/src/router.get.ts
@@ -1,10 +1,8 @@
 import path from 'node:path';
-import { server } from './route-server.js';
 import { ConsoleLog } from './lib/Console.js';
+import { server } from './route-server.js';
 
-export async function get(req: Request): Promise<void | Response> {
-  const url = new URL(req.url);
-  const pathname = decodeURIComponent(url.pathname);
+export async function get(req: Request, url: URL, pathname: string): Promise<void | Response> {
   ConsoleLog(`GET      ${pathname}`);
 
   // server api