+++ dev_server/tools/all.ts
@@ -1,2 +1,0 @@
-Bun.spawnSync(['bun', 'update']);
-Bun.spawnSync(['bun', 'run', 'format']);