+++ src/lib/ericchase/Algorithm/Stream/Compare.ts
@@ -17,36 +17,43 @@ export class U8StreamReader {
     this.value = value;
     return { changed: true };
   }
+  releaseLock() {
+    this.reader.releaseLock();
+  }
 }
 
 export async function U8StreamCompare(stream1: ReadableStream<Uint8Array>, stream2: ReadableStream<Uint8Array>) {
   const one = new U8StreamReader(stream1.getReader());
   const two = new U8StreamReader(stream2.getReader());
-
-  while (true) {
-    let changed = false;
-    if (one.done === false && one.i >= one.length) {
-      if ((await one.next()).changed === true) {
-        changed = true;
+  try {
+    while (true) {
+      let changed = false;
+      if (one.done === false && one.i >= one.length) {
+        if ((await one.next()).changed === true) {
+          changed = true;
+        }
       }
-    }
-    if (two.done === false && two.i >= two.length) {
-      if ((await two.next()).changed === true) {
-        changed = true;
+      if (two.done === false && two.i >= two.length) {
+        if ((await two.next()).changed === true) {
+          changed = true;
+        }
       }
-    }
-    if (one.done && two.done) {
-      return true;
-    }
-    if (one.done !== two.done || changed === false) {
-      return false;
-    }
-    while (one.i < one.length && two.i < two.length) {
-      if (one.value[one.i] !== two.value[two.i]) {
+      if (one.done && two.done) {
+        return true;
+      }
+      if (one.done !== two.done || changed === false) {
         return false;
       }
-      one.i++;
-      two.i++;
+      while (one.i < one.length && two.i < two.length) {
+        if (one.value[one.i] !== two.value[two.i]) {
+          return false;
+        }
+        one.i++;
+        two.i++;
+      }
     }
+  } finally {
+    one.releaseLock();
+    two.releaseLock();
   }
 }