+++ src/lib/ericchase/Algorithm/Stream/ReadAll.ts
@@ -2,15 +2,17 @@ import { U8Concat } from '../Array/Uint8Array.js';
 
 export async function U8StreamReadAll(stream: ReadableStream<Uint8Array>): Promise<Uint8Array> {
   const reader = stream.getReader();
-  const chunks: Uint8Array[] = [];
-  while (true) {
-    const { done, value } = await reader.read();
-    if (value) {
+  try {
+    const chunks: Uint8Array[] = [];
+    while (true) {
+      const { done, value } = await reader.read();
+      if (done) {
+        break;
+      }
       chunks.push(value);
     }
-    if (done) {
-      break;
-    }
+    return U8Concat(chunks);
+  } finally {
+    reader.releaseLock();
   }
-  return U8Concat(chunks);
 }