+++ tools/docker-db.ts
@@ -1,8 +1,8 @@
-import { ConsoleLog } from '../src/lib/ericchase/Utility/Console.js';
+import { ConsoleError } from '../src/lib/ericchase/Utility/Console.js';
 
 const [command] = Bun.argv.slice(2);
 
-ConsoleLog('reminder: make sure docker desktop is running');
+ConsoleError('reminder: make sure docker desktop is running');
 if (command === 'start') {
   Bun.spawnSync(['docker-compose', 'up', '-d'], { cwd: './dev_database', stdout: 'inherit', stderr: 'inherit' });
 }