+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+### 2024-09-13
+
+**dev_server**
+
+- add basic hot reloading!
+- minor updates to route handlers
+
+**src/lib/ericchase**
+
+- fix potential bug with releasing stream locks
+  - /Algorithm/Stream/
+- add AsyncReader classes for ReadableStreams
+  - /Algorithm/Stream/AsyncReader.ts
+- add PathManager class for working with paths
+  - /Platform/Bun/Path.ts
+- rework PathGroup class to better handle extensions
+- add tests for PathGroup and PathManager
+
+**tools**
+
+- add file cache to support incremental builds!
+- add BuildRunner class for better bundling with Bun
+- update html preprocessor system to use classes
+- add new html preprocessor to change script tag src extensions
+- rework build and dev scripts for cleaner more efficient flow
+- change default sourcemap mode from inline to linked
+
+will update this later if i missed anything
+
 ### 2024-09-11
 
 **dev_database**